move list item's font styles into file with item spacing, rather than inside outerItemStyles
