+++ src/components/Header/SideNav/outerItemStyles.js
@@ -1,11 +1,4 @@
 import { css } from 'styled-components'
-import remcalc from 'remcalc'
-
-export const fontSizeAndWeight = css`
-  font-weight: 400;
-  font-size: ${remcalc(21)};
-  line-height: ${remcalc(24)};
-`
 
 export const defaultStyles = css`
   color: ${props => props.theme.colors.white};
