+++ src/components/Header/SideNav/sideNavItemStyles.js
@@ -1,9 +1,13 @@
 import { css } from 'styled-components'
 import remcalc from 'remcalc'
 
-const sideNavItemSpacing = css`
+export const fontSizeAndWeight = css`
+  font-weight: 400;
+  font-size: ${remcalc(21)};
+  line-height: ${remcalc(24)};
+`
+
+export const itemSpacing = css`
   padding: ${remcalc('9')} ${remcalc('34')} ${remcalc('7')} ${remcalc('20')};
   margin: ${remcalc('4')};
 `
-
-export default sideNavItemSpacing
